+++ algorithms/flybyPoint/flybyPointAlgorithm.h
@@ -29,36 +29,35 @@ class FlybyPointAlgorithm final {
     std::tuple<Eigen::Vector3d, Eigen::Vector3d, Eigen::Vector3d> computeGuidanceSolution() const;
     double getTimeBetweenFilterData() const;
     void setTimeBetweenFilterData(double timeBetweenFilterData);
-    double getToleranceForCollinearity() const;
-    void setToleranceForCollinearity(double toleranceForCollinearity);
+    float getToleranceForCollinearity() const;
+    void setToleranceForCollinearity(float toleranceForCollinearity);
     int getSignOfOrbitNormalFrameVector() const;
     void setSignOfOrbitNormalFrameVector(int signOfOrbitNormalFrameVector);
-    double getMaximumAccelerationThreshold() const;
-    void setMaximumAccelerationThreshold(double maxAccelerationThreshold);
-    double getMaximumRateThreshold() const;
-    void setMaximumRateThreshold(double maxRateThreshold);
-    double getPositionKnowledgeSigma() const;
-    void setPositionKnowledgeSigma(double positionKnowledgeStd);
+    float getMaximumAccelerationThreshold() const;
+    void setMaximumAccelerationThreshold(float maxAccelerationThreshold);
+    float getMaximumRateThreshold() const;
+    void setMaximumRateThreshold(float maxRateThreshold);
+    float getPositionKnowledgeSigma() const;
+    void setPositionKnowledgeSigma(float positionKnowledgeStd);
 
    private:
-    double dt = 0;                     //!< current time step between last two updates
-    double timeOfFirstRead = 0;        //!< time of first nav solution read
-    double timeBetweenFilterData = 0;  //!< time between two subsequent reads of the filter information
-    double toleranceForCollinearity =
-        0;  //!< tolerance for singular conditions when position and velocity are collinear
+    double dt = 0;                       //!< current time step between last two updates
+    double timeOfFirstRead = 0;          //!< time of first nav solution read
+    double timeBetweenFilterData = 0;    //!< time between two subsequent reads of the filter information
+    float toleranceForCollinearity = 0;  //!< tolerance for singular conditions when position and velocity are collinear
     int signOfOrbitNormalFrameVector = 1;  //!< Sign of orbit normal vector to complete reference frame
 
-    double maxRate = 0;          //!< maximum rate spacecraft can control to, used for validity of solution
-    double maxAcceleration = 0;  //!< maximum acceleration spacecraft can control to, used for validity of solution
+    float maxRate = 0;          //!< maximum rate spacecraft can control to, used for validity of solution
+    float maxAcceleration = 0;  //!< maximum acceleration spacecraft can control to, used for validity of solution
 
     bool firstRead = true;            //!< variable to attest if this is the first read after a Reset
     double f0 = 0;                    //!< ratio between relative velocity and position norms at time of read [Hz]
     double gamma0 = 0;                //!< flight path angle of the spacecraft at time of read [rad]
     uint64_t lastFilterReadTime = 0;  //!< time of last filter read
-    Eigen::Matrix3d R0N{Eigen::Matrix3d::Identity()};            //!< inertial-to-reference DCM at time of read
+    Eigen::Matrix3f R0N{Eigen::Matrix3f::Identity()};            //!< inertial-to-reference DCM at time of read
     Eigen::Vector3d firstNavPosition = Eigen::Vector3d::Zero();  //!< First position used to create profile
     Eigen::Vector3d firstNavVelocity = Eigen::Vector3d::Zero();  //!< First velocity used to create profile
-    double positionKnowledgeSigma = 0;                           //!< Last position used to create profile
+    float positionKnowledgeSigma = 0;                            //!< Last position used to create profile
 };
 
 #endif  // F32XMERA_FLYBY_POINT_ALGORITHM_H
